backend/pyproject.toml
[project]
name = "pga-backend"
version = "0.1.0"
description = "Policy Gap Analysis Backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart",
    "pydantic>=2.0",
    "pydantic-settings",
    "supabase>=2.12.0",
    "python-jose[cryptography]",
    # Neo4j & Qdrant integration
    "neo4j>=5.15.0",
    "qdrant-client>=1.7.0",
    "openai>=1.12.0",
    "tiktoken>=0.5.0",
    "tenacity>=8.2.0",
    # Web crawling & AI agents
    "crawl4ai>=0.4.0",
    "playwright>=1.40.0",
    "anthropic>=0.42.0",
    # Document extraction
    "llama-cloud-services>=0.1.0",
    # Fallback text extraction (no API keys needed)
    "pypdf>=4.0.0",
    "python-docx>=1.0.0",
    "openpyxl>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v"

[dependency-groups]
dev = [
    "ruff>=0.14.14",
]
